Extract some components and make card selection a little nicer visually
