+++ frontend/src/Components/ResultsPage/compareVotes.test.ts
