+++ frontend/src/types/WebSocket.ts
@@ -1,7 +1,7 @@
 export type CardValue =
   | 'not-voted'
   | '?'
-  | 'coffee'
+  | '☕️'
   | '0'
   | '0.5'
   | '1'
